flatten utils
