+++ packages/insomnia-inso/src/db/models/util.ts
@@ -1,4 +1,4 @@
-import { InsoError } from '../../errors';
+import { InsoError } from '../../util';
 import type { BaseModel } from './types';
 
 export const matchIdIsh = ({ _id }: BaseModel, identifier: string) =>
