+++ packages/insomnia-inso/src/logger.ts
@@ -1,25 +1,16 @@
 import consola, { BasicReporter, FancyReporter, LogLevel, logType } from 'consola';
-
-import { noop } from './util';
-
+export const noop = () => { };
 export const noConsoleLog = async <T>(callback: () => Promise<T>): Promise<T> => {
   const oldConsoleLog = console.log;
-
   console.log = noop;
-
   try {
     return await callback();
   } finally {
     console.log = oldConsoleLog;
   }
 };
-
-export type LogsByType = {
-  [t in logType]?: string[]
-};
-
-export type ModifiedConsola = ReturnType<typeof consola.create> & { __getLogs: () => LogsByType};
-
+export type LogsByType = { [t in logType]?: string[] };
+export type ModifiedConsola = ReturnType<typeof consola.create> & { __getLogs: () => LogsByType };
 const consolaLogger = consola.create({
   reporters: [
     new FancyReporter({
@@ -32,12 +23,9 @@ const consolaLogger = consola.create({
 });
 
 (consolaLogger as ModifiedConsola).__getLogs = () => ({});
-
 export const logger = consolaLogger as ModifiedConsola;
-
 export const configureLogger = (verbose = false, ci = false) => {
   logger.level = verbose ? LogLevel.Verbose : LogLevel.Info;
-
   if (ci) {
     logger.setReporters([new BasicReporter()]);
   }
